requirements.txt
asgiref==3.7.2
blessed==1.20.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
curtsies==0.4.1
cwcwidth==0.1.8
distlib==0.3.6
Django==4.2.3
django-appconf==1.0.5
django-browser-reload==1.10.0
django-compressor==4.4
django-contrib-comments==2.2.0
django-debug-toolbar==4.1.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-render-block==0.9.2
django-rich==1.6.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==18.11.2
filelock==3.12.2
greenlet==2.0.2
gunicorn==21.2.0
identify==2.5.24
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.5
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
Pillow==10.0.0
platformdirs==3.8.0
podman-compose==1.0.6
pre-commit==3.3.3
Pygments==2.15.1
pynvim==0.4.3
pytailwindcss==0.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyxdg==0.28
PyYAML==6.0
rcssmin==1.1.1
requests==2.31.0
rich==13.4.2
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
urllib3==2.0.3
virtualenv==20.23.1
wcwidth==0.2.6
